+++ core/src/main/java/es/urjc/etsii/grafo/solver/executors/WorkUnit.java
@@ -5,5 +5,5 @@ import es.urjc.etsii.grafo.solution.Solution;
 import es.urjc.etsii.grafo.solver.algorithms.Algorithm;
 import es.urjc.etsii.grafo.solver.services.ExceptionHandler;
 
-public record WorkUnit<S extends Solution<S,I>, I extends Instance>(String experimentName, String instanceName, Algorithm<S, I> algorithm, int i, ExceptionHandler<S,I> exceptionHandler) {
+public record WorkUnit<S extends Solution<S,I>, I extends Instance>(String experimentName, String instancePath, Algorithm<S, I> algorithm, int i, ExceptionHandler<S,I> exceptionHandler) {
 }