+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # v0.12
 - (New) GRASP::afterGRASP: Optionally execute anything after the GRASP constructive finishes
 - (New) Verify move changes solution score as expected
-- (Fix) Export all generated solutions by default if a solution serializer is enabled instead of only best for pair (alg, instance) 
+- (New) Option to enable/disable instance preload
+- (New) Warn if IteratedImproved appears stuck
+- (Refactor) Results Serializers: CSV, XLSX
+- (Fix) Export all generated solutions by default if a solution serializer is enabled instead of only best for pair (alg, instance)
+
+
 
 # v0.11
 - (New) Swap methods in ArrayUtil